requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
WTForms==3.2.1
Pillow==11.1.0
python-dotenv==1.0.1
email-validator==2.2.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
razorpay==1.4.2
